requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-multipart
pandas
python-dotenv
psycopg2-binary
